[RC driver] Allow low-level key remapping from Python
Allow the Python code to set key remapping in the C++ eRCInput code.

Currently only effective in eRCDeviceInputDev. Remapping can be readily added to any other eRCDevice subclass that requires it.

In eRCDeviceInputDev::handleCode() supplied remappings over-ride the compiled-in remappings, but if the Python code supplies no remappings (the default state, or if the Python code explicitly sets the mapping to an empty dict), the compiled-in remappings apply.

That maintains backwards-compatibility.