[ConfigList.py] Add a settings save notifier
This change adds a new facility to allow Setup sub class based code to be notified when the user saves changed settings.

This can be considered an alternative to setting notifiers on a number of individual settings.  It is particularly helpful if a number of different settings contribute to a single required change.  When using separate notifiers the same change code could be called multiple times.  With this simple notifier call back system a Setup sub class can register to be notified if nominated settings have been changed and then perform a required action once.

The following methods have been added:
- addSaveNotifier()
- removeSaveNotifier()
- clearSaveNotifiers()
The methods do as their names imply.

An example of how this change is intended to be used will soon be made to Time.py and NetworkTime.py.