Refactor MultiBoot
This change refactors the entire MultiBoot reference tool library in Tools.  This allowed for refactoring, simplifying and improving of the MultiBoot functionality and UI.
- The new MultiBoot code can now uses the "enigma.info" file to determine the contents of a slot.
- If "enigma.info" is not available the code will fall back to using the "/etc/issue" file and the "opkg" status to determine the contents of the slot as happened previously.
- The new MultiBoot tool code will now also provide information about newer slot usages like Android, AndroidLinuxSE and Recovery slots.
- The code has been written to no longer import or use SystemInfo.py.  This change eliminates the previous risk of creating boot loops if the order of imports is not precise.
- All external commands are now directly executed and no longer require the instantiation of shells to do the work.
- The MultiBoot data is now cached for faster access.  The data can be refreshed from disk by calling the "loadMultiBoot()" method.
- A number of methods have been provided in the tool library to extract slot, boot mode and other slot information.  (Please use these methods rather than groping around the code for internal variables.)
- The MultiBoot tool module now handles all aspects of slot management.  Any future need to manage slots should be added to this module only so that all Enigma2 code can find and use this common slot management interface.

MultiBootSelector.py has been replaced with MultiBootManager.  The new module allows users to select a boot slot, as before, but users can now also delete, restore and wipe slots as appropriate for the slot.  This new module is designed to purely be a user interface to the MultiBoot library code, it no longer directly manipulates the slots itself.

SystemInfo.py has been updated to use the new MultiBoot tool library.  The MultiBoot import has been moved to the top of the file as the import order is no longer significant or relevant.  The MultiBoot variables have been trimmed down to a minimum to maintain code compatibility.  Code can now consult MultiBoot directly for any information required about the slots.

Thanks go to Captain, Jbleyel and nautilus7 for their assistance in developing, proofing and testing the new code.