[SleepTimer.py] Add button mode
This change allows the user to press SLEEP, or assign another button to act as SLEEP, to bring up an interactive version of the SleepTimer screen where only the Sleep Timer options are shown and the user may press OK to select and save the Sleep Timer options.

The user interface for this screen when called from a menu is unchanged.